fix: silence -Wsign-compare and -Wimplicit-fallthrough on Pi 2 build
GCC 12.2 (armv7 / Pi 2) flagged three warnings on the post-Phase-13
tree. All are benign, but we keep the warning noise floor at zero.

rds.c: FILTER_SIZE was `sizeof(waveform_biphase)/sizeof(float)`, which
has type size_t. Three loop counters and one index in
rds_get_samples() are `int`, which provoked -Wsign-compare on every
comparison. Fix: cast FILTER_SIZE to int at the macro level; the
actual value (576) fits comfortably. SAMPLE_BUFFER_SIZE (= 768)
follows automatically.

fm_mpx.c: the sample-loop counter in fm_mpx_get_samples() was `int`
but `length` is `size_t` (static in this TU). Fix: make the loop
counter `size_t`.

pi_fm_rds.c: the ':' case in the getopt_long_only switch calls
fatal() and relies on it not returning; GCC cannot see that without
an attribute, so it warns about fall-through into 'case '?''. Fix:
add __attribute__((noreturn)) to fatal's forward declaration. The
format(printf,...) attribute is deliberately NOT added because some
callers use glibc's "%m" extension, which would then warn.

Build now completes warning-free on Pi 2 (-Wall -Wextra -std=gnu99).

Made-with: Cursor

+++ src/fm_mpx.c
@@ -166,7 +166,7 @@ pifm_status_t fm_mpx_get_samples(float *mpx_buffer) {
 
     if(inf  == NULL) return PIFM_OK; // if there is no audio, stop here
     
-    for(int i=0; i<length; i++) {
+    for(size_t i=0; i<length; i++) {
         if(audio_pos >= downsample_factor) {
             audio_pos -= downsample_factor;
